+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Master
+* Add reset-choice command.
+
 ## 1.1.0
 * Migrate to lsp-protocol.
 * Implement metals/windowStateDidChange notification.